requirements.txt
Django>=4.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
redis>=5.0.1
celery>=5.3.6
gunicorn>=21.2.0
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file support
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
